Move controller to admin area
