+++ src/Submissions/Api/Areas/Admin/Views/PagesAdmin/ContentPage.cshtml
@@ -1,6 +1,5 @@
-﻿@using Biobanks.Submissions.Api.Models.Submissions
-@using Biobanks.Submissions.Api.HtmlHelpers
-@model ContentPageModel
+﻿@using Biobanks.Submissions.Api.HtmlHelpers
+@model Biobanks.Submissions.Api.Areas.Admin.Models.PagesAdmin.ContentPageModel
 
 @{
     ViewBag.Title = Model.Title;
