+++ src/Submissions/Api/Areas/Admin/Views/PagesAdmin/CreateEdit.cshtml
@@ -1,5 +1,4 @@
-﻿@using Biobanks.Submissions.Api.Models.Submissions
-@model ContentPageModel
+﻿@model Biobanks.Submissions.Api.Areas.Admin.Models.PagesAdmin.ContentPageModel
 @{
     ViewBag.Title = "Create & Edit Content Page";
 }
