+++ src/Submissions/Api/Areas/Admin/Views/PagesAdmin/CreateEditPreview.cshtml
@@ -1,5 +1,4 @@
-﻿@using Biobanks.Submissions.Api.Models.Submissions
-@model ContentPageModel
+﻿@model Biobanks.Submissions.Api.Areas.Admin.Models.PagesAdmin.ContentPageModel
 
 <style>
     .navbar-fixed-top {
