+++ src/Submissions/Api/Areas/Admin/Views/PagesAdmin/Index.cshtml
@@ -1,5 +1,4 @@
-﻿@using Biobanks.Submissions.Api.Models.Submissions
-@model IEnumerable<ContentPageModel>
+﻿@model IEnumerable<Biobanks.Submissions.Api.Areas.Admin.Models.PagesAdmin.ContentPageModel>
 
 @{
 	ViewBag.Title = "Content Page Admin";
